+++ polynomial_symmetry/polynomial_symmetry.js
@@ -5,8 +5,8 @@ import { DemoShader } from './shaders/DemoShader.js';
 import { RosetteCurveShader } from './shaders/RosetteCurveShader.js';
 import { Checkerboard, HalfPlanes, WebcamTexture } from './Texture.js';
 import { Coefficients } from './Coefficients.js';
-import { PointSymmetry } from './SymmetryRule.js';
-import { MAX_TERMS, TWO_PI, mod } from './util.js';
+import { PointSymmetry } from './PointSymmetry.js';
+import { MAX_TERMS, TWO_PI, mod } from './math_util.js';
 
 import './components/Checkbox.js';
 import './components/Dropdown.js';
